+++ system/libraries/Javascript.php
@@ -637,14 +637,13 @@ class CI_Javascript {
 		}
 		elseif ($this->CI->config->item('javascript_location') !== '')
 		{
-			$this->_javascript_location = 
-				$this->CI->config->item('javascript_location');
+			$this->_javascript_location = $this->CI->config->item('javascript_location');
 		}
 
 		
 		// If Relative OR Arg is a complete url then load the file from the Arg
-		if ($relative === TRUE OR strpos($external_file, 'http://') === 0 || 
-								  strpos($external_file, 'https://') === 0 ||
+		if ($relative === TRUE OR strpos($external_file, 'http://') === 0 OR 
+								  strpos($external_file, 'https://') === 0 OR
 								  strpos($external_file, '//') === 0 )
 		{
 			$str = $this->_open_script($external_file);
@@ -652,8 +651,8 @@ class CI_Javascript {
 		
 		// Does the config have a complete url?  "//" is for CDNs that 
 		// support both.
-		elseif (strpos($this->_javascript_location, 'http://') === 0 ||
-				strpos($this->_javascript_location, "https://") === 0 ||
+		elseif (strpos($this->_javascript_location, 'http://') === 0 OR
+				strpos($this->_javascript_location, "https://") === 0 OR
 				strpos($this->_javascript_location, "//") === 0 )
 		{
 			$str = $this->_open_script($this->_javascript_location.$external_file);
@@ -662,8 +661,7 @@ class CI_Javascript {
 		// Last chance base it off of our base url
 		else
 		{
-			$str = $this->_open_script($this->CI->config->slash_item('base_url').
-						$this->_javascript_location.$external_file);
+			$str = $this->_open_script($this->CI->config->slash_item('base_url') . $this->_javascript_location.$external_file);
 		}
 
 		return $str.$this->_close_script();