+++ user_guide_src/source/changelog.rst
@@ -254,7 +254,10 @@ Release Date: Not Released
 	 -  Added support for timestamp-based migrations (enabled by default).
 	 -  Added ``$config['migration_type']`` to allow switching between *sequential* and *timestamp* migrations.
    -  :doc:`User Agent Library <libraries/user_agent>` will now check if robots are pretending to be mobile clients (helps with e.g. Google indexing mobile website versions).
-
+   - :doc:`Javascript Library <libraries/javascript>` changes includes:
+   	-  Added support for protocol-less urls.
+   
+   
 -  Core
 
    -  :doc:`URI Library <libraries/uri>` changes include: